requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioop-lts==0.2.2
backoff==2.2.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
fastapi==0.124.0
frozenlist==1.8.0
fsspec==2024.12.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==0.3.80
langchain-mongodb==0.9.0
langchain-neo4j==0.4.0
langchain-openai==1.1.0
langchain-qdrant==1.1.0
langchain-text-splitters==0.3.11
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-mongodb==0.3.0
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
lark==1.3.1
marshmallow==3.26.1
mcp==1.24.0
mem0ai==1.0.1
multidict==6.7.0
mypy_extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.6.1
numpy==2.3.5
openai==2.9.0
openai-agents==0.6.3
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
portalocker==3.2.0
posthog==7.0.1
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.5
pymongo-search-utils==0.1.0
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.1
rank-bm25==0.2.2
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rq==2.6.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.4
SQLAlchemy==2.0.44
sse-starlette==3.0.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uuid_utils==0.12.0
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
